Loops and assignments and correct ordering sorta

+++ RapidsLang/Parser/Nodes/LiteralNumberNode.cs
@@ -2,4 +2,4 @@ using RapidsLang.Lexer;
 
 namespace RapidsLang.Parser.Nodes;
 
-public record LiteralNumberNode(Token Token, float Number) : ExpressionNode;
+public record LiteralNumberNode(Token Token, double Number) : ExpressionNode;
